Moved template for blog posts to templates folder
